requirements.txt
discord.py
toml
sqlalchemy